+++ client/src/main/scala/org/http4s/client/middleware/Retry.scala
@@ -2,16 +2,19 @@ package org.http4s
 package client
 package middleware
 
+import java.time.Instant
+import java.time.temporal.ChronoUnit
+
+import cats._
+import cats.implicits._
+import fs2._
+import org.http4s.Status._
+import org.http4s.headers.`Retry-After`
+import org.log4s.getLogger
+
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
-import scala.math.{pow, min, random}
-import org.http4s.Status._
-import org.log4s.getLogger
-import fs2.{Scheduler, Strategy, Task}
-
-import scala.Either
-import scala.Right
-import scala.Left
+import scala.math.{min, pow, random}
 
 object Retry {
 
@@ -32,11 +35,11 @@ object Retry {
     def prepareLoop(req: Request, attempts: Int): Task[DisposableResponse] = {
       client.open(req).attempt flatMap {
         // TODO fs2 port - Reimplement request isIdempotent in some form
-        case Right(dr @ DisposableResponse(Response(status, _, _, _, _), _)) if RetriableStatuses(status) =>
+        case Right(dr @ DisposableResponse(Response(status, _, headers, _, _), _)) if RetriableStatuses(status) =>
           backoff(attempts) match {
             case Some(duration) =>
               logger.info(s"Request ${req} has failed on attempt #${attempts} with reason ${status}. Retrying after ${duration}.")
-              dr.dispose.flatMap(_ => nextAttempt(req, attempts, duration))
+              dr.dispose.flatMap(_ => nextAttempt(req, attempts, duration, headers.get(`Retry-After`)))
             case None =>
               logger.info(s"Request ${req} has failed on attempt #${attempts} with reason ${status}. Giving up.")
               Task.now(dr)
@@ -47,7 +50,7 @@ object Retry {
           backoff(attempts) match {
             case Some(duration) =>
               logger.error(e)(s"Request ${req} threw an exception on attempt #${attempts} attempts. Retrying after ${duration}.")
-              nextAttempt(req, attempts, duration)
+              nextAttempt(req, attempts, duration, None)
             case None =>
               // info instead of error(e), because e is not discarded
               logger.info(s"Request ${req} threw an exception on attempt #${attempts} attempts. Giving up.")
@@ -56,9 +59,15 @@ object Retry {
       }
     }
 
-    def nextAttempt(req: Request, attempts: Int, duration: FiniteDuration): Task[DisposableResponse] = {
-      // TODO honor Retry-After header
-      prepareLoop(req.withEmptyBody, attempts + 1).schedule(duration)
+    def nextAttempt(req: Request, attempts: Int, duration: FiniteDuration, retryHeader: Option[`Retry-After`]): Task[DisposableResponse] = {
+      val headerDuration = retryHeader.map { h =>
+        h.retry match {
+          case Left(d) => Instant.now().until(d.toInstant, ChronoUnit.SECONDS)
+          case Right(secs) => secs
+        }
+      }.getOrElse(0L)
+      val sleepDuration = Math.max(headerDuration, duration.length).seconds
+      prepareLoop(req.withEmptyBody, attempts + 1).schedule(sleepDuration)
     }
 
     client.copy(open = Service.lift(prepareLoop(_, 1)))